+++ tests/test_binapy.py
@@ -9,33 +9,26 @@ BINARY = b"\xd2m'\x10\x7f\xa9\xb0\xf4f\x9e\x85\xedBK%\x93"
 BASE64 = b"0m0nEH+psPRmnoXtQkslkw=="
 BASE64U = b"0m0nEH-psPRmnoXtQkslkw"
 SHA256 = b"\x91\x02\xef\xd5U\xa6\xefr\xc9B\x18\x16\xa9\xf2\x8d\xf1Ur\x1d\xd1\x9en\xa0)\xf8\xb6\x0b\x93\x81\xce\xbc="
-
 JSON = {"foo": "bar"}
 
 
-@pytest.fixture
-def binary_sample():
-    return BINARY
-
-
-def test_binary(binary_sample):
-    bp = BinaPy(binary_sample)
+def test_binapy() -> None:
+    bp = BinaPy(BINARY)
     assert len(bp) == 16
-    b64 = bp.encode_b64()
+    b64 = bp.encode_to("b64")
     assert isinstance(b64, BinaPy)
     assert b64 == BASE64
-    assert bp.encode_b64u() == BASE64U
-    sha256 = bp.hash_sha256()
+    assert bp.encode_to("b64u") == BASE64U
+    sha256 = bp.encode_to("sha256")
     assert isinstance(sha256, BinaPy)
     assert sha256 == SHA256
-    assert bp.hash_sha256().encode_b64() == BinaPy(SHA256).encode_b64()
+    assert bp.encode_to("sha256").encode_to("b64") == BinaPy(SHA256).encode_to("b64")
     assert isinstance(bp[:], BinaPy)
     assert isinstance(bp[4:-2], BinaPy)
 
-    assert not bp.is_b64()
-    assert BinaPy(BASE64).is_b64()
-
     assert not bp.check("b64")
+    assert BinaPy(BASE64).check("b64")
+
     b64 = BinaPy(BASE64)
     assert b64.check("b64")
 
@@ -59,14 +52,14 @@ def test_binary(binary_sample):
     assert BinaPy.serialize_from("json", JSON).parse_to("json") == JSON
 
 
-def test_helloworld():
-    hello_world = "Hello, World!"
-    bp = BinaPy(hello_world).compress_gzip().encode_b64u()
+def test_helloworld() -> None:
+    hello_world = b"Hello, World!"
+    bp = BinaPy(hello_world).encode_to("gzip").encode_to("b64u")
     assert bp == b"eJzzSM3JyddRCM8vyklRBAAfngRq"
-    assert bp.decode_b64u().decompress_gzip().decode() == hello_world
+    assert bp.decode_from("b64u").decode_from("gzip") == hello_world
 
 
-def test_int():
+def test_int() -> None:
     i = 0x12345678901234567890
     bp = BinaPy.from_int(i)
     assert bp == b"\x12\x34\x56\x78\x90\x12\x34\x56\x78\x90"
@@ -77,13 +70,13 @@ def test_int():
     assert bp.to_int() == i
 
 
-def test_random():
+def test_random() -> None:
     bp = BinaPy.random(12)
     assert len(bp) == 12
     assert bp != b"\00" * 12
 
 
-def test_unknown_features():
+def test_unknown_features() -> None:
     bp = BinaPy.random(12)
     with pytest.raises(ValueError):
         bp.encode_to("something_not_known")
@@ -101,9 +94,9 @@ def test_unknown_features():
         BinaPy.serialize_from("something_not_known", {"foo": "bar"})
 
 
-def test_exceptions():
+def test_exceptions() -> None:
     @binapy_decoder("some_feature")
-    def check_some_feature(bp):
+    def decode_some_feature(bp: BinaPy) -> BinaPy:
         return bp
 
     bp = BinaPy()
@@ -112,7 +105,7 @@ def test_exceptions():
     assert bp.check("some_feature", decode=True) is True
 
     @binapy_decoder("other_feature")
-    def check_other_feature(bp):
+    def check_other_feature(bp: BinaPy) -> bool:
         raise ValueError()
 
     bp = BinaPy()